Conclusion: Create Code
We have been commissioned to create a hotel management application. We need a web portal that allows people to book an available room. We also need a desktop app for when they check in.
We need to create a minimum viable product (MVP) that will allow the customer to get going and to identify what features should be added later.